requirements.txt
python-binance
streamlit
python-dotenv